requirements.txt
torch>=2.3
numpy>=1.24
pandas>=2.1
pyyaml>=6.0
typer>=0.12
rich>=13.7
tqdm>=4.66
matplotlib>=3.8
seaborn>=0.13
pytest>=8.1
pytest-cov>=5.0
ruamel.yaml>=0.18
